Refactored code to be in different files for each component
