+++ src/App.css
@@ -9,6 +9,7 @@ button.checkout-button {
 button.checkout-button-display {
   display: block;
 }
+
 .App-header {
   background-color: #222;
   height: 150px;
