+++ src/App.js
@@ -3,6 +3,14 @@ import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import './App.css';
 import axios from 'axios'
 
+import Search from './components/Search'
+import Home from './components/Home'
+import Header from './components/Header'
+import Customers from './components/Customers'
+import Customer from './components/Customer'
+import Library from './components/Library'
+import Movie from './components/Movie'
+
 class App extends Component {
   constructor() {
     super();
@@ -35,14 +43,15 @@ class App extends Component {
 
     .then((response) => {
       console.log(response)
+      this.setState(
+        {
+          movieTitle: "", 
+          customer: "",
+          checkoutButtonClassName: "checkout-button",
+      })
     })
 
-    this.setState(
-      {
-        movieTitle: "", 
-        customer: "",
-        checkoutButtonClassName: "checkout-button",
-    })
+    
     .catch((error)=>{
       console.log(error.message)
       this.setState({error: error.message, errorClass: 'display-error'});
@@ -103,207 +112,6 @@ class App extends Component {
       </Router>
     );
   }
-}
-
-function Home() {
-  return <h2>Home</h2>;
-}
-
-class Search extends Component {
-  constructor() {
-    super();
-    this.state = {
-      title: "",
-      searchList: []
-    }
-  }
-
-  onChangeTitle = (event) => {
-    let title = this.state.title;
-    // const field = event.target.name;
-    const value = event.target.value;
-    title = value;
-    this.setState({title})
-
-  }
-
-  onFormSubmit = (event) => {
-    event.preventDefault();
-    axios.get('http://localhost:3090/movies?query=' + this.state.title.toString())
-    .then((response) => {
-      console.log(response.data)
-      const searchList = response.data.map((movie) => {
-          return movie
-      })
-      this.setState({searchList})
-    })
-    
-  }
-
-
-  onMovieSelect = (movie) => {
-    return () => {
-    axios.post('http://localhost:3090/movies', 
-    {
-      title: movie.title,
-      overview: movie.overview,
-      release_date: movie.release_date,
-      image_url: movie.image_url,
-      external_id: movie.external_id,
-    }
-    )
-    .then((response) => {
-      console.log(response)
-      // be submitting post request to video store api from response?
-    })
-      let newState = this.state
-      newState.searchList = [];
-      this.setState({newState});
-    }
-  }
-
-  searchDisplay = () => {
-    return this.state.searchList.map((movie) => {
-        return (
-          <div>
-            <p>{movie.title}</p>
-            <p onClick={this.onMovieSelect(movie)}>Select!</p>
-          </div>
-        )
-    })
-  }
-
-  render () {
-    return (
-      <div>
-        <h2>Search</h2>
-        <form onSubmit={this.onFormSubmit}>
-          <label htmlFor="title">Title</label>
-          <input name="title" type="text" value={this.state.title} onChange={this.onChangeTitle}/>
-          <input type="submit" value="Search" />
-        </form>
-        <h4>{this.searchDisplay()}</h4>
-      </div>
-    );
-  }
-}
-
-function Movie (props) {
-  const onTitleClick = () => {
-    props.addMovieToRentCallback(props.title)
-  }
-
-  return <p onClick={onTitleClick}>{props.title}</p>
-}
-
-class Library extends Component {
-  constructor(props) {
-    super(props)
-  
-    this.state = {
-      movieList: [],
-    };
-  }
-
-    generateMovieList = () => {
-      console.log(this.props)
-      return this.state.movieList.map((movie) => {
-        return (<Movie 
-        key={movie.id}
-        id={movie.id}
-        title={movie.title}
-        addMovieToRentCallback = {this.props.addMovieToRentCallback}
-        />)
-      })
-    }
-
-    componentDidMount() { 
-      axios.get("http://localhost:3090/movies") 
-      .then((response)=>{
-        const movieList = response.data.map((movie) => {
-          return movie
-        })
-      this.setState({movieList})
-    })
-  }
-
-
-render () {
-  return (
-  <div>
-  <h2>Library</h2>
-  <h2>{this.generateMovieList()}</h2>
-  </div>
-  );
-  }
-}
-
-function Customer (props) {
-  const onCustomerClick = () => {
-    props.addCustomertoRentCallback(props.customer)
-  }
-  return <p onClick={onCustomerClick}>{props.customer.name}</p>
-}
-
-class Customers extends Component {
-  constructor() {
-    super();
-    
-    this.state = {
-      customerList: [],
-    };
-  }
-
-  generateCustomerList = () => {
-    return this.state.customerList.map((customer) => {
-      return (<Customer 
-        key={customer.id}
-        customer={customer}
-        addCustomertoRentCallback = {this.props.addCustomerToRentCallback}
-      />)
-    })
-  }
-
-  componentDidMount() { 
-    axios.get("http://localhost:3090/customers") 
-    .then((response)=>{
-    const customerList = response.data.map((customer) => {
-      return customer
-    })
-    console.log(customerList)
-    this.setState({customerList})
-
-  })
-
-}
- 
-  render () {
-    return (
-    <div>
-      <h2>{this.generateCustomerList()}</h2>
-    </div>
-  );
-}
-}
-
-function Header() {
-  return (
-    <ul>
-      <li>
-        <Link to="/">Home</Link>
-      </li>
-      <li>
-        <Link to="/search">Search</Link>
-      </li>
-      <li>
-        <Link to="/library">Library</Link>
-      </li>
-
-      <li>
-        <Link to="/customers">Customers</Link>
-      </li>
-    </ul>
-  );
 }
 
 export default App;
