+++ src/components/Header.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { BrowserRouter as Router, Route, Link } from "react-router-dom";
+
+function Header() {
+  return (
+    <ul>
+      <li>
+        <Link to="/">Home</Link>
+      </li>
+      <li>
+        <Link to="/search">Search</Link>
+      </li>
+      <li>
+        <Link to="/library">Library</Link>
+      </li>
+
+      <li>
+        <Link to="/customers">Customers</Link>
+      </li>
+    </ul>
+  );
+}
+
+export default Header;
