+++ src/components/Home.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Home.css';
+
+function Home() {
+  return <h2>Home</h2>;
+}
+
+export default Home;
