+++ src/components/Movie.js
@@ -1,0 +1,12 @@
+import React from 'react';
+// import './Home.css';
+
+function Movie (props) {
+  const onTitleClick = () => {
+    props.addMovieToRentCallback(props.title)
+  }
+
+  return <p onClick={onTitleClick}>{props.title}</p>
+}
+
+export default Movie;
